+++ azure-functions-language-worker-protobuf/src/proto/identity/ClaimsIdentityRpc.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 // protobuf vscode extension: https://marketplace.visualstudio.com/items?itemName=zxh404.vscode-proto3
 
-import "shared/NullableString.proto";
+import "shared/NullableTypes.proto";
 
 // Light-weight representation of a .NET System.Security.Claims.ClaimsIdentity object.
 // This is the same serialization as found in EasyAuth, and needs to be kept in sync with
