+++ types/public/Interfaces.d.ts
@@ -1,3 +1,4 @@
+/// <reference types="long" />
 /**
  * Interface for your Azure Function code. This function must be exported (via module.exports or exports)
  * and will execute when triggered. It is recommended that you declare this function as async, which
@@ -110,6 +111,32 @@ export interface HttpRequest {
  * Possible values for an HTTP request method.
  */
 export declare type HttpMethod = "GET" | "POST" | "DELETE" | "HEAD" | "PATCH" | "PUT" | "OPTIONS" | "TRACE" | "CONNECT";
+/**
+ * Http response cookie object to "Set-Cookie"
+ */
+export interface Cookie {
+    /** Cookie name */
+    name: string;
+    /** Cookie value */
+    value: string;
+    /** Specifies allowed hosts to receive the cookie */
+    domain?: string;
+    /** Specifies URL path that must exist in the requested URL */
+    path?: string;
+    /**
+     * Sets the cookie to expire at a specific date instead of when the client closes.
+     * It is generally recommended that you use maxAge over expires.
+     */
+    expires?: Date | number;
+    /** Sets the cookie to only be sent with an encrypted request */
+    secure?: boolean;
+    /** Sets the cookie to be inaccessible to JavaScript's Document.cookie API */
+    httpOnly?: boolean;
+    /** Can restrict the cookie to not be sent with cross-site requests */
+    sameSite?: "Strict" | "Lax" | undefined;
+    /** Number of seconds until the cookie expires. A zero or negative number will expire the cookie immediately. */
+    maxAge?: Long;
+}
 export interface ExecutionContext {
     /**
      * A unique GUID per function invocation.
